+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Setup Bundler default group only when using as an API to avoid loading :development/:test group gems into a production app 
 - Announce that download files are missing for `bundler download show` when a Downloadfile exists but downloads have not been made
+- Document options (e.g. --keep-existing and --all-operating-systems) in `bundle download help`
 
 ## 1.2.0